gs-aiflow/app/requirements.txt
flask==2.2.2
gevent==22.10.1
flask_restful==0.3.9
flask_sockets==0.2.1
flask_cors==3.0.10
pyyaml==6.0
mysql-connector-python==8.0.31
kubernetes==25.3.0
apscheduler==3.10.1
notebook==6.5.4